camera: move Camera_Initialise to TRX
