+++ src/tr1/game/camera/common.h
@@ -4,6 +4,5 @@
 
 #include <stdint.h>
 
-void Camera_Initialise(void);
 void Camera_UpdateCutscene(void);
 void Camera_MoveManual(void);
